fix cache not working

+++ includes/modules/cache/RapidLoad_Cache_Store.php
@@ -392,7 +392,7 @@ class RapidLoad_Cache_Store
 
     public static function cache_expired( $cache_file ) {
 
-        if ( RapidLoad_Cache_Engine::$settings['cache_expiry_time'] === 0 ) {
+        if ( ((int)RapidLoad_Cache_Engine::$settings['cache_expiry_time']) === 0 ) {
             return false;
         }
